OCA-Management-System/OCA_Management_System/requirements.txt
Flask
Flask-SQLAlchemy
mysql.connector.python